@@ -4,7 +4,7 @@ version: 3.4.0
 homepage: https://github.com/Ali-Azmoud/flutter_xlider
 
 environment:
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
