@@ -14,48 +14,48 @@ import 'package:flutter/material.dart';
 import 'dart:math';
 
 class FlutterSlider extends StatefulWidget {
-  final Key key;
+  final Key? key;
   final Axis axis;
-  final double handlerWidth;
-  final double handlerHeight;
-  final FlutterSliderHandler handler;
-  final FlutterSliderHandler rightHandler;
-  final Function(int handlerIndex, dynamic lowerValue, dynamic upperValue)
+  final double? handlerWidth;
+  final double? handlerHeight;
+  final FlutterSliderHandler? handler;
+  final FlutterSliderHandler? rightHandler;
+  final Function(int handlerIndex, dynamic lowerValue, dynamic upperValue)?
       onDragStarted;
-  final Function(int handlerIndex, dynamic lowerValue, dynamic upperValue)
+  final Function(int handlerIndex, dynamic lowerValue, dynamic upperValue)?
       onDragCompleted;
-  final Function(int handlerIndex, dynamic lowerValue, dynamic upperValue)
+  final Function(int handlerIndex, dynamic lowerValue, dynamic upperValue)?
       onDragging;
-  final double min;
-  final double max;
+  final double? min;
+  final double? max;
   final List<double> values;
-  final List<FlutterSliderFixedValue> fixedValues;
+  final List<FlutterSliderFixedValue>? fixedValues;
   final bool rangeSlider;
   final bool rtl;
   final bool jump;
   final bool selectByTap;
   final List<FlutterSliderIgnoreSteps> ignoreSteps;
   final bool disabled;
-  final double touchSize;
+  final double? touchSize;
   final bool visibleTouchArea;
   final double minimumDistance;
   final double maximumDistance;
   final FlutterSliderHandlerAnimation handlerAnimation;
-  final FlutterSliderTooltip tooltip;
+  final FlutterSliderTooltip? tooltip;
   final FlutterSliderTrackBar trackBar;
   final FlutterSliderStep step;
-  final FlutterSliderHatchMark hatchMark;
+  final FlutterSliderHatchMark? hatchMark;
   final bool centeredOrigin;
   final bool lockHandlers;
-  final double lockDistance;
-  final BoxDecoration decoration;
-  final BoxDecoration foregroundDecoration;
+  final double? lockDistance;
+  final BoxDecoration? decoration;
+  final BoxDecoration? foregroundDecoration;
 
   FlutterSlider(
       {this.key,
       this.min,
       this.max,
-      @required this.values,
+      required this.values,
       this.fixedValues,
       this.axis = Axis.horizontal,
       this.handler,
@@ -139,25 +139,25 @@ class _FlutterSliderState extends State<FlutterSlider>
     with TickerProviderStateMixin {
   bool __isInitCall = true;
 
-  double _touchSize;
+  double? _touchSize;
 
-  Widget leftHandler;
-  Widget rightHandler;
+  Widget? leftHandler;
+  Widget? rightHandler;
 
-  double _leftHandlerXPosition = 0;
-  double _rightHandlerXPosition = 0;
-  double _leftHandlerYPosition = 0;
-  double _rightHandlerYPosition = 0;
+  double? _leftHandlerXPosition = 0;
+  double? _rightHandlerXPosition = 0;
+  double? _leftHandlerYPosition = 0;
+  double? _rightHandlerYPosition = 0;
 
-  double _lowerValue = 0;
-  double _upperValue = 0;
+  double? _lowerValue = 0;
+  double? _upperValue = 0;
   dynamic _outputLowerValue = 0;
   dynamic _outputUpperValue = 0;
 
-  double _realMin;
-  double _realMax;
+  double? _realMin;
+  double? _realMax;
 
-  double _divisions;
+  late double _divisions;
   double _handlersPadding = 0;
 
   GlobalKey leftHandlerKey = GlobalKey();
@@ -166,49 +166,49 @@ class _FlutterSliderState extends State<FlutterSlider>
   GlobalKey leftTooltipKey = GlobalKey();
   GlobalKey rightTooltipKey = GlobalKey();
 
-  double _handlersWidth;
-  double _handlersHeight;
+  double? _handlersWidth;
+  double? _handlersHeight;
 
-  double _constraintMaxWidth;
-  double _constraintMaxHeight;
+  late double _constraintMaxWidth;
+  late double _constraintMaxHeight;
 
-  double _containerWidthWithoutPadding;
-  double _containerHeightWithoutPadding;
+  double? _containerWidthWithoutPadding;
+  double? _containerHeightWithoutPadding;
 
   double _containerLeft = 0;
   double _containerTop = 0;
 
-  FlutterSliderTooltip _tooltipData;
+  late FlutterSliderTooltip _tooltipData;
 
-  List<Function> _positionedItems;
+  late List<Function> _positionedItems;
 
   double _rightTooltipOpacity = 0;
   double _leftTooltipOpacity = 0;
 
-  AnimationController _rightTooltipAnimationController;
-  Animation<Offset> _rightTooltipAnimation;
-  AnimationController _leftTooltipAnimationController;
-  Animation<Offset> _leftTooltipAnimation;
+  late AnimationController _rightTooltipAnimationController;
+  Animation<Offset>? _rightTooltipAnimation;
+  late AnimationController _leftTooltipAnimationController;
+  Animation<Offset>? _leftTooltipAnimation;
 
-  AnimationController _leftHandlerScaleAnimationController;
-  Animation<double> _leftHandlerScaleAnimation;
-  AnimationController _rightHandlerScaleAnimationController;
-  Animation<double> _rightHandlerScaleAnimation;
+  AnimationController? _leftHandlerScaleAnimationController;
+  Animation<double>? _leftHandlerScaleAnimation;
+  AnimationController? _rightHandlerScaleAnimationController;
+  Animation<double>? _rightHandlerScaleAnimation;
 
-  double _containerHeight;
-  double _containerWidth;
+  double? _containerHeight;
+  double? _containerWidth;
 
   int _decimalScale = 0;
 
   double xDragTmp = 0;
   double yDragTmp = 0;
 
-  double xDragStart;
-  double yDragStart;
+  double? xDragStart;
+  double? yDragStart;
 
-  double _widgetStep;
-  double _widgetMin;
-  double _widgetMax;
+  double? _widgetStep;
+  double? _widgetMin;
+  double? _widgetMax;
   List<FlutterSliderIgnoreSteps> _ignoreSteps = [];
   List<FlutterSliderFixedValue> _fixedValues = [];
 
@@ -216,7 +216,7 @@ class _FlutterSliderState extends State<FlutterSlider>
 
   bool __dragging = false;
 
-  double __dAxis,
+  double? __dAxis,
       __rAxis,
       __axisDragTmp,
       __axisPosTmp,
@@ -225,10 +225,10 @@ class _FlutterSliderState extends State<FlutterSlider>
       __leftHandlerPosition,
       __containerSizeWithoutHalfPadding;
 
-  Orientation oldOrientation;
+  Orientation? oldOrientation;
 
   double __lockedHandlersDragOffset = 0;
-  double _distanceFromRightHandler, _distanceFromLeftHandler;
+  double? _distanceFromRightHandler, _distanceFromLeftHandler;
   double _handlersDistance = 0;
 
   bool _slidingByActiveTrackBar = false;
@@ -255,8 +255,8 @@ class _FlutterSliderState extends State<FlutterSlider>
   void dispose() {
     _rightTooltipAnimationController.dispose();
     _leftTooltipAnimationController.dispose();
-    _leftHandlerScaleAnimationController.dispose();
-    _rightHandlerScaleAnimationController.dispose();
+    _leftHandlerScaleAnimationController!.dispose();
+    _rightHandlerScaleAnimationController!.dispose();
     super.dispose();
   }
 
@@ -271,18 +271,19 @@ class _FlutterSliderState extends State<FlutterSlider>
           _constraintMaxWidth = constraints.maxWidth;
           _constraintMaxHeight = constraints.maxHeight;
 
-          _containerWidthWithoutPadding = _constraintMaxWidth - _handlersWidth;
+          _containerWidthWithoutPadding = _constraintMaxWidth - _handlersWidth!;
           _containerHeightWithoutPadding =
-              _constraintMaxHeight - _handlersHeight;
+              _constraintMaxHeight - _handlersHeight!;
 
-          double sliderProperSize = _findProperSliderSize();
+          double? sliderProperSize = _findProperSliderSize();
           if (widget.axis == Axis.vertical) {
             double layoutWidth = constraints.maxWidth;
             if (layoutWidth == double.infinity) {
               layoutWidth = 0;
             }
             __containerSizeWithoutPadding = _containerHeightWithoutPadding;
-            _containerWidth = [(sliderProperSize * 2), layoutWidth].reduce(max);
+            _containerWidth =
+                [(sliderProperSize! * 2), layoutWidth].reduce(max);
             _containerHeight = constraints.maxHeight;
           } else {
             double layoutHeight = constraints.maxHeight;
@@ -291,7 +292,7 @@ class _FlutterSliderState extends State<FlutterSlider>
             }
             _containerWidth = constraints.maxWidth;
             _containerHeight =
-                [(sliderProperSize * 2), layoutHeight].reduce(max);
+                [(sliderProperSize! * 2), layoutHeight].reduce(max);
             __containerSizeWithoutPadding = _containerWidthWithoutPadding;
           }
 
@@ -332,15 +333,15 @@ class _FlutterSliderState extends State<FlutterSlider>
     );
   }
 
-  double _findProperSliderSize() {
+  double? _findProperSliderSize() {
     List<double> sizes = [
       widget.trackBar.activeTrackBarHeight,
       widget.trackBar.inactiveTrackBarHeight
     ];
     if (widget.axis == Axis.horizontal) {
-      sizes.add(_handlersHeight);
+      sizes.add(_handlersHeight ?? 0);
     } else {
-      sizes.add(_handlersWidth);
+      sizes.add(_handlersWidth ?? 0);
     }
 
     return sizes.reduce(max);
@@ -375,13 +376,13 @@ class _FlutterSliderState extends State<FlutterSlider>
     _leftHandlerScaleAnimation =
         Tween(begin: 1.0, end: widget.handlerAnimation.scale).animate(
             CurvedAnimation(
-                parent: _leftHandlerScaleAnimationController,
+                parent: _leftHandlerScaleAnimationController!,
                 reverseCurve: widget.handlerAnimation.reverseCurve,
                 curve: widget.handlerAnimation.curve));
     _rightHandlerScaleAnimation =
         Tween(begin: 1.0, end: widget.handlerAnimation.scale).animate(
             CurvedAnimation(
-                parent: _rightHandlerScaleAnimationController,
+                parent: _rightHandlerScaleAnimationController!,
                 reverseCurve: widget.handlerAnimation.reverseCurve,
                 curve: widget.handlerAnimation.curve));
 
@@ -391,24 +392,24 @@ class _FlutterSliderState extends State<FlutterSlider>
     if (widget.rangeSlider == true &&
         widget.maximumDistance != null &&
         widget.maximumDistance > 0 &&
-        (_upperValue - _lowerValue) > widget.maximumDistance) {
+        (_upperValue! - _lowerValue!) > widget.maximumDistance) {
       throw 'lower and upper distance is more than maximum distance';
     }
     if (widget.rangeSlider == true &&
         widget.minimumDistance != null &&
         widget.minimumDistance > 0 &&
-        (_upperValue - _lowerValue) < widget.minimumDistance) {
+        (_upperValue! - _lowerValue!) < widget.minimumDistance) {
       throw 'lower and upper distance is less than minimum distance';
     }
 
     Offset animationStart = Offset(0, 0);
     if (widget.tooltip != null &&
-        widget.tooltip.disableAnimation != null &&
-        widget.tooltip.disableAnimation) {
+        widget.tooltip!.disableAnimation != null &&
+        widget.tooltip!.disableAnimation!) {
       animationStart = Offset(0, -1);
     }
 
-    Offset animationFinish;
+    Offset? animationFinish;
     switch (_tooltipData.direction) {
       case FlutterSliderTooltipDirection.top:
         animationFinish = Offset(0, -1);
@@ -430,7 +431,7 @@ class _FlutterSliderState extends State<FlutterSlider>
       _rightTooltipAnimationController = AnimationController(
           duration: Duration(milliseconds: 200), vsync: this);
     } else {
-      if (_tooltipData.alwaysShowTooltip) {
+      if (_tooltipData.alwaysShowTooltip!) {
         _rightTooltipOpacity = _leftTooltipOpacity = 1;
       }
     }
@@ -447,7 +448,7 @@ class _FlutterSliderState extends State<FlutterSlider>
                 parent: _rightTooltipAnimationController,
                 curve: Curves.fastOutSlowIn));
 
-    WidgetsBinding.instance.addPostFrameCallback((_) {
+    WidgetsBinding.instance!.addPostFrameCallback((_) {
       _renderBoxInitialization();
 
       _arrangeHandlersPosition();
@@ -459,7 +460,7 @@ class _FlutterSliderState extends State<FlutterSlider>
   }
 
   void _drawHatchMark() {
-    if (widget.hatchMark == null || widget.hatchMark.disabled) return;
+    if (widget.hatchMark == null || widget.hatchMark!.disabled) return;
     _points = [];
 
     double maxTrackBarHeight = [
@@ -468,40 +469,40 @@ class _FlutterSliderState extends State<FlutterSlider>
     ].reduce(max);
 
     FlutterSliderHatchMark hatchMark = FlutterSliderHatchMark();
-    hatchMark.disabled = widget.hatchMark.disabled ?? false;
-    hatchMark.density = widget.hatchMark.density ?? 1;
+    hatchMark.disabled = widget.hatchMark!.disabled ?? false;
+    hatchMark.density = widget.hatchMark!.density ?? 1;
     hatchMark.linesDistanceFromTrackBar =
-        widget.hatchMark.linesDistanceFromTrackBar ?? 0;
+        widget.hatchMark!.linesDistanceFromTrackBar ?? 0;
     hatchMark.labelsDistanceFromTrackBar =
-        widget.hatchMark.labelsDistanceFromTrackBar ?? 0;
-    hatchMark.smallLine = widget.hatchMark.smallLine ??
+        widget.hatchMark!.labelsDistanceFromTrackBar ?? 0;
+    hatchMark.smallLine = widget.hatchMark!.smallLine ??
         FlutterSliderSizedBox(
             height: 5,
             width: 1,
             decoration: BoxDecoration(color: Colors.black45));
-    hatchMark.bigLine = widget.hatchMark.bigLine ??
+    hatchMark.bigLine = widget.hatchMark!.bigLine ??
         FlutterSliderSizedBox(
             height: 9,
             width: 2,
             decoration: BoxDecoration(color: Colors.black45));
-    hatchMark.labelBox = widget.hatchMark.labelBox ??
+    hatchMark.labelBox = widget.hatchMark!.labelBox ??
         FlutterSliderSizedBox(height: 50, width: 50);
-    hatchMark.labels = widget.hatchMark.labels ?? null;
-    hatchMark.linesAlignment = widget.hatchMark.linesAlignment ??
+    hatchMark.labels = widget.hatchMark!.labels ?? null;
+    hatchMark.linesAlignment = widget.hatchMark!.linesAlignment ??
         FlutterSliderHatchMarkAlignment.right;
-    hatchMark.displayLines = widget.hatchMark.displayLines ?? false;
+    hatchMark.displayLines = widget.hatchMark!.displayLines ?? false;
 
-    if (hatchMark.displayLines) {
+    if (hatchMark.displayLines!) {
       double percent = 100 * hatchMark.density;
       double barWidth, barHeight, distance;
-      double linesTop, linesLeft, linesRight, linesBottom;
+      double? linesTop, linesLeft, linesRight, linesBottom;
 
       if (widget.axis == Axis.horizontal) {
 //      top = hatchMark.linesDistanceFromTrackBar - 2.25;
-        distance = ((_constraintMaxWidth - _handlersWidth) / percent);
+        distance = ((_constraintMaxWidth - _handlersWidth!) / percent);
       } else {
 //      left = hatchMark.linesDistanceFromTrackBar - 3.62;
-        distance = ((_constraintMaxHeight - _handlersHeight) / percent);
+        distance = ((_constraintMaxHeight - _handlersHeight!) / percent);
       }
 
       Alignment linesAlignment;
@@ -521,17 +522,17 @@ class _FlutterSliderState extends State<FlutterSlider>
 
       Widget barLine;
       for (int p = 0; p <= percent; p++) {
-        FlutterSliderSizedBox barLineBox = hatchMark.smallLine;
+        FlutterSliderSizedBox? barLineBox = hatchMark.smallLine;
 
         if (p % 5 - 1 == -1) {
           barLineBox = hatchMark.bigLine;
         }
 
         if (widget.axis == Axis.horizontal) {
-          barHeight = barLineBox.height;
+          barHeight = barLineBox!.height;
           barWidth = barLineBox.width;
         } else {
-          barHeight = barLineBox.width;
+          barHeight = barLineBox!.width;
           barWidth = barLineBox.height;
         }
 
@@ -551,30 +552,30 @@ class _FlutterSliderState extends State<FlutterSlider>
           linesLeft = (p * distance) + _handlersPadding - 0.75;
           if (hatchMark.linesAlignment ==
               FlutterSliderHatchMarkAlignment.right) {
-            linesTop = _containerHeight / 2 + maxTrackBarHeight / 2 + 2;
-            linesBottom = _containerHeight / 2 - maxTrackBarHeight - 15;
+            linesTop = _containerHeight! / 2 + maxTrackBarHeight / 2 + 2;
+            linesBottom = _containerHeight! / 2 - maxTrackBarHeight - 15;
           } else {
-            linesTop = _containerHeight / 2 - maxTrackBarHeight - 15;
-            linesBottom = _containerHeight / 2 + maxTrackBarHeight / 2 + 2;
+            linesTop = _containerHeight! / 2 - maxTrackBarHeight - 15;
+            linesBottom = _containerHeight! / 2 + maxTrackBarHeight / 2 + 2;
           }
           if (hatchMark.linesAlignment == FlutterSliderHatchMarkAlignment.left)
-            linesBottom += hatchMark.linesDistanceFromTrackBar;
+            linesBottom += hatchMark.linesDistanceFromTrackBar!;
           else
-            linesTop += hatchMark.linesDistanceFromTrackBar;
+            linesTop += hatchMark.linesDistanceFromTrackBar!;
         } else {
           linesTop = (p * distance) + _handlersPadding - 0.5;
           if (hatchMark.linesAlignment ==
               FlutterSliderHatchMarkAlignment.right) {
-            linesLeft = _containerWidth / 2 + maxTrackBarHeight / 2 + 2;
-            linesRight = _containerWidth / 2 - maxTrackBarHeight - 15;
+            linesLeft = _containerWidth! / 2 + maxTrackBarHeight / 2 + 2;
+            linesRight = _containerWidth! / 2 - maxTrackBarHeight - 15;
           } else {
-            linesLeft = _containerWidth / 2 - maxTrackBarHeight - 15;
-            linesRight = _containerWidth / 2 + maxTrackBarHeight / 2 + 2;
+            linesLeft = _containerWidth! / 2 - maxTrackBarHeight - 15;
+            linesRight = _containerWidth! / 2 + maxTrackBarHeight / 2 + 2;
           }
           if (hatchMark.linesAlignment == FlutterSliderHatchMarkAlignment.left)
-            linesRight += hatchMark.linesDistanceFromTrackBar;
+            linesRight += hatchMark.linesDistanceFromTrackBar!;
           else
-            linesLeft += hatchMark.linesDistanceFromTrackBar;
+            linesLeft += hatchMark.linesDistanceFromTrackBar!;
         }
 
         _points.add(Positioned(
@@ -586,35 +587,36 @@ class _FlutterSliderState extends State<FlutterSlider>
       }
     }
 
-    if (hatchMark.labels != null && hatchMark.labels.length > 0) {
+    if (hatchMark.labels != null && hatchMark.labels!.length > 0) {
       List<Widget> labelWidget = [];
-      Widget label;
+      Widget? label;
       double labelBoxHalfSize;
-      double top, left, bottom, right;
-      double tr;
-      for (FlutterSliderHatchMarkLabel markLabel in hatchMark.labels) {
+      double? top, left, bottom, right;
+      double? tr;
+      for (FlutterSliderHatchMarkLabel markLabel in hatchMark.labels!) {
         label = markLabel.label;
         tr = markLabel.percent;
         labelBoxHalfSize = 0;
 
-        if (widget.rtl) tr = 100 - tr;
+        if (widget.rtl) tr = 100 - tr!;
 
         if (widget.axis == Axis.horizontal) {
-          labelBoxHalfSize = hatchMark.labelBox.width / 2 - 0.5;
+          labelBoxHalfSize = hatchMark.labelBox!.width / 2 - 0.5;
         } else {
-          labelBoxHalfSize = hatchMark.labelBox.height / 2 - 0.5;
+          labelBoxHalfSize = hatchMark.labelBox!.height / 2 - 0.5;
         }
 
         labelWidget = [
           Container(
-            height:
-                widget.axis == Axis.vertical ? hatchMark.labelBox.height : null,
-            width: widget.axis == Axis.horizontal
-                ? hatchMark.labelBox.width
+            height: widget.axis == Axis.vertical
+                ? hatchMark.labelBox!.height
                 : null,
-            decoration: hatchMark.labelBox.decoration,
-            foregroundDecoration: hatchMark.labelBox.foregroundDecoration,
-            transform: hatchMark.labelBox.transform,
+            width: widget.axis == Axis.horizontal
+                ? hatchMark.labelBox!.width
+                : null,
+            decoration: hatchMark.labelBox!.decoration,
+            foregroundDecoration: hatchMark.labelBox!.foregroundDecoration,
+            transform: hatchMark.labelBox!.transform,
             child: Align(alignment: Alignment.center, child: label),
           )
         ];
@@ -625,7 +627,7 @@ class _FlutterSliderState extends State<FlutterSlider>
             mainAxisAlignment: MainAxisAlignment.center,
             children: labelWidget,
           );
-          left = tr * _containerWidthWithoutPadding / 100 -
+          left = tr! * _containerWidthWithoutPadding! / 100 -
               0.5 +
               _handlersPadding -
               labelBoxHalfSize;
@@ -638,7 +640,7 @@ class _FlutterSliderState extends State<FlutterSlider>
             mainAxisAlignment: MainAxisAlignment.center,
             children: labelWidget,
           );
-          top = tr * _containerHeightWithoutPadding / 100 -
+          top = tr! * _containerHeightWithoutPadding! / 100 -
               0.5 +
               _handlersPadding -
               labelBoxHalfSize;
@@ -657,11 +659,11 @@ class _FlutterSliderState extends State<FlutterSlider>
       throw 'when range mode is true, slider needs both lower and upper values';
 
     if (widget.fixedValues == null) {
-      if (widget.values[0] != null && widget.values[0] < _widgetMin)
+      if (widget.values[0] != null && widget.values[0] < _widgetMin!)
         throw 'Lower value should be greater than min';
 
       if (widget.rangeSlider == true) {
-        if (widget.values[1] != null && widget.values[1] > _widgetMax)
+        if (widget.values[1] != null && widget.values[1] > _widgetMax!)
           throw 'Upper value should be smaller than max';
       }
     } else {
@@ -693,7 +695,7 @@ class _FlutterSliderState extends State<FlutterSlider>
 
     _ignoreSteps = [];
 
-    if (widget.fixedValues != null && widget.fixedValues.length > 0) {
+    if (widget.fixedValues != null && widget.fixedValues!.length > 0) {
       _realMax = 100;
       _realMin = 0;
       _widgetStep = 1;
@@ -701,8 +703,8 @@ class _FlutterSliderState extends State<FlutterSlider>
       _widgetMin = 0;
 
       List<double> fixedValuesIndices = [];
-      for (FlutterSliderFixedValue fixedValue in widget.fixedValues) {
-        fixedValuesIndices.add(fixedValue.percent.toDouble());
+      for (FlutterSliderFixedValue fixedValue in widget.fixedValues!) {
+        fixedValuesIndices.add(fixedValue.percent!.toDouble());
       }
 
       double lowerIgnoreBound = -1;
@@ -710,9 +712,9 @@ class _FlutterSliderState extends State<FlutterSlider>
       List<double> fixedV = [];
       for (double fixedPercent = 0; fixedPercent <= 100; fixedPercent++) {
         dynamic fValue = '';
-        for (FlutterSliderFixedValue fixedValue in widget.fixedValues) {
+        for (FlutterSliderFixedValue fixedValue in widget.fixedValues!) {
           if (fixedValue.percent == fixedPercent.toInt()) {
-            fixedValuesIndices.add(fixedValue.percent.toDouble());
+            fixedValuesIndices.add(fixedValue.percent!.toDouble());
             fValue = fixedValue.value;
 
             upperIgnoreBound = fixedPercent;
@@ -733,14 +735,14 @@ class _FlutterSliderState extends State<FlutterSlider>
         }
       }
 
-      double biggestPoint =
+      double? biggestPoint =
           _findBiggestIgnorePoint(ignoreBeyondBoundaries: true);
       if (!fixedV.contains(100)) {
         _ignoreSteps
-            .add(FlutterSliderIgnoreSteps(from: biggestPoint + 1, to: 101));
+            .add(FlutterSliderIgnoreSteps(from: biggestPoint! + 1, to: 101));
       }
     } else {
-      _realMax = _widgetMax - _widgetMin;
+      _realMax = _widgetMax! - _widgetMin!;
       _widgetStep = widget.step.step;
     }
 
@@ -784,11 +786,11 @@ class _FlutterSliderState extends State<FlutterSlider>
 
     _generateHandler();
 
-    _handlersDistance = widget.lockDistance ?? _upperValue - _lowerValue;
+    _handlersDistance = widget.lockDistance ?? _upperValue! - _lowerValue!;
   }
 
   void _setDivisionAndDecimalScale() {
-    _divisions = _realMax / _widgetStep;
+    _divisions = _realMax! / _widgetStep!;
     String tmpDecimalScale = '0';
     List<String> tmpDecimalScaleArr = _widgetStep.toString().split(".");
     if (tmpDecimalScaleArr.length > 1) tmpDecimalScale = tmpDecimalScaleArr[1];
@@ -797,8 +799,8 @@ class _FlutterSliderState extends State<FlutterSlider>
     }
   }
 
-  List<double> _calculateUpperAndLowerValues() {
-    double localLV, localUV;
+  List<double?> _calculateUpperAndLowerValues() {
+    double? localLV, localUV;
     localLV = widget.values[0];
     if (widget.rangeSlider) {
       localUV = widget.values[1];
@@ -819,10 +821,10 @@ class _FlutterSliderState extends State<FlutterSlider>
   }
 
   void _setValues() {
-    List<double> localValues = _calculateUpperAndLowerValues();
+    List<double?> localValues = _calculateUpperAndLowerValues();
 
-    _lowerValue = localValues[0] - _widgetMin;
-    _upperValue = localValues[1] - _widgetMin;
+    _lowerValue = localValues[0]! - _widgetMin!;
+    _upperValue = localValues[1]! - _widgetMin!;
 
     _outputUpperValue = _displayRealValue(_upperValue);
     _outputLowerValue = _displayRealValue(_lowerValue);
@@ -831,8 +833,8 @@ class _FlutterSliderState extends State<FlutterSlider>
       _outputLowerValue = _displayRealValue(_upperValue);
       _outputUpperValue = _displayRealValue(_lowerValue);
 
-      double tmpUpperValue = _realMax - _lowerValue;
-      double tmpLowerValue = _realMax - _upperValue;
+      double tmpUpperValue = _realMax! - _lowerValue!;
+      double tmpLowerValue = _realMax! - _upperValue!;
 
       _lowerValue = tmpLowerValue;
       _upperValue = tmpUpperValue;
@@ -842,11 +844,11 @@ class _FlutterSliderState extends State<FlutterSlider>
   void _arrangeHandlersPosition() {
     if (!__dragging) {
       if (widget.axis == Axis.horizontal) {
-        _handlersPadding = _handlersWidth / 2;
+        _handlersPadding = _handlersWidth! / 2;
         _leftHandlerXPosition = getPositionByValue(_lowerValue);
         _rightHandlerXPosition = getPositionByValue(_upperValue);
       } else {
-        _handlersPadding = _handlersHeight / 2;
+        _handlersPadding = _handlersHeight! / 2;
         _leftHandlerYPosition = getPositionByValue(_lowerValue);
         _rightHandlerYPosition = getPositionByValue(_upperValue);
       }
@@ -901,28 +903,28 @@ class _FlutterSliderState extends State<FlutterSlider>
 
   double getPositionByValue(value) {
     if (widget.axis == Axis.horizontal)
-      return (((_constraintMaxWidth - _handlersWidth) / _realMax) * value) -
-          (_touchSize);
+      return (((_constraintMaxWidth - _handlersWidth!) / _realMax!) * value) -
+          _touchSize!;
     else
-      return (((_constraintMaxHeight - _handlersHeight) / _realMax) * value) -
-          (_touchSize);
+      return (((_constraintMaxHeight - _handlersHeight!) / _realMax!) * value) -
+          _touchSize!;
   }
 
   double getValueByPosition(double position) {
-    double value = ((position / (__containerSizeWithoutPadding / _divisions)) *
-        _widgetStep);
+    double value = ((position / (__containerSizeWithoutPadding! / _divisions)) *
+        _widgetStep!);
     value = (double.parse(value.toStringAsFixed(_decimalScale)) -
-        double.parse((value % _widgetStep).toStringAsFixed(_decimalScale)));
+        double.parse((value % _widgetStep!).toStringAsFixed(_decimalScale)));
     return value;
   }
 
-  double getLengthByValue(value) {
+  double? getLengthByValue(value) {
     return value * __containerSizeWithoutPadding / _realMax;
   }
 
   double getValueByPositionIgnoreOffset(double position) {
-    double value = ((position / (__containerSizeWithoutPadding / _divisions)) *
-        _widgetStep);
+    double value = ((position / (__containerSizeWithoutPadding! / _divisions)) *
+        _widgetStep!);
     return value;
   }
 
@@ -930,10 +932,10 @@ class _FlutterSliderState extends State<FlutterSlider>
       {double lockedHandlersDragOffset = 0,
       double tappedPositionWithPadding = 0,
       bool selectedByTap = false}) {
-    if (widget.disabled || (widget.handler != null && widget.handler.disabled))
+    if (widget.disabled || (widget.handler != null && widget.handler!.disabled))
       return;
 
-    _handlersDistance = widget.lockDistance ?? _upperValue - _lowerValue;
+    _handlersDistance = widget.lockDistance ?? _upperValue! - _lowerValue!;
 
     // Tip: lockedHandlersDragOffset only subtracts from left handler position
     // because it calculates drag position only by left handler's position
@@ -965,17 +967,17 @@ class _FlutterSliderState extends State<FlutterSlider>
       __leftHandlerPosition = _leftHandlerYPosition;
     }
 
-    __axisPosTmp = __dAxis - __axisDragTmp + (_touchSize);
+    __axisPosTmp = __dAxis! - __axisDragTmp! + _touchSize!;
 
-    _checkRangeStep(getValueByPositionIgnoreOffset(__axisPosTmp));
+    _checkRangeStep(getValueByPositionIgnoreOffset(__axisPosTmp!));
 
-    __rAxis = getValueByPosition(__axisPosTmp);
+    __rAxis = getValueByPosition(__axisPosTmp!);
 
     if (widget.rangeSlider &&
         widget.minimumDistance > 0 &&
-        (__rAxis + widget.minimumDistance) >= _upperValue) {
-      _lowerValue = (_upperValue - widget.minimumDistance > _realMin)
-          ? _upperValue - widget.minimumDistance
+        (__rAxis! + widget.minimumDistance) >= _upperValue!) {
+      _lowerValue = (_upperValue! - widget.minimumDistance > _realMin!)
+          ? _upperValue! - widget.minimumDistance
           : _realMin;
       _updateLowerValue(_lowerValue);
 
@@ -984,50 +986,51 @@ class _FlutterSliderState extends State<FlutterSlider>
 
     if (widget.rangeSlider &&
         widget.maximumDistance > 0 &&
-        __rAxis <= (_upperValue - widget.maximumDistance)) {
-      _lowerValue = (_upperValue - widget.maximumDistance > _realMin)
-          ? _upperValue - widget.maximumDistance
+        __rAxis! <= (_upperValue! - widget.maximumDistance)) {
+      _lowerValue = (_upperValue! - widget.maximumDistance > _realMin!)
+          ? _upperValue! - widget.maximumDistance
           : _realMin;
       _updateLowerValue(_lowerValue);
 
       if (lockedHandlersDragOffset == 0) validMove = validMove & false;
     }
 
-    double tS = _touchSize;
+    double? tS = _touchSize;
     if (widget.jump) {
-      tS = _touchSize + _handlersPadding;
+      tS = _touchSize! + _handlersPadding;
     }
 
     validMove = validMove & _leftHandlerIgnoreSteps(tS);
 
     bool forcePosStop = false;
-    if (((__axisPosTmp <= 0) ||
-        (__axisPosTmp - tS >= __rightHandlerPosition))) {
+    if (((__axisPosTmp! <= 0) ||
+        (__axisPosTmp! - tS! >= __rightHandlerPosition!))) {
       forcePosStop = true;
     }
 
     if (validMove &&
-        ((__axisPosTmp + _handlersPadding >= _handlersPadding) ||
+        ((__axisPosTmp! + _handlersPadding >= _handlersPadding) ||
             forcePosStop)) {
-      double tmpLowerValue = __rAxis;
+      double tmpLowerValue = __rAxis!;
 
-      if (tmpLowerValue > _realMax) tmpLowerValue = _realMax;
-      if (tmpLowerValue < _realMin) tmpLowerValue = _realMin;
+      if (tmpLowerValue > _realMax!) tmpLowerValue = _realMax!;
+      if (tmpLowerValue < _realMin!) tmpLowerValue = _realMin!;
 
-      if (tmpLowerValue > _upperValue) tmpLowerValue = _upperValue;
+      if (tmpLowerValue > _upperValue!) tmpLowerValue = _upperValue!;
 
       if (widget.jump == true) {
         if (!forcePosStop) {
           _lowerValue = tmpLowerValue;
-          _leftHandlerMoveBetweenSteps(__dAxis - __axisDragTmp, selectedByTap);
+          _leftHandlerMoveBetweenSteps(
+              __dAxis! - __axisDragTmp!, selectedByTap);
           __leftHandlerPosition = getPositionByValue(_lowerValue);
         } else {
-          if (__axisPosTmp - tS >= __rightHandlerPosition) {
+          if (__axisPosTmp! - tS! >= __rightHandlerPosition!) {
             __leftHandlerPosition = __rightHandlerPosition;
-            _lowerValue = tmpLowerValue = _upperValue;
+            _lowerValue = tmpLowerValue = _upperValue!;
           } else {
             __leftHandlerPosition = getPositionByValue(_realMin);
-            _lowerValue = tmpLowerValue = _realMin;
+            _lowerValue = tmpLowerValue = _realMin!;
           }
           _updateLowerValue(tmpLowerValue);
         }
@@ -1035,17 +1038,17 @@ class _FlutterSliderState extends State<FlutterSlider>
         _lowerValue = tmpLowerValue;
 
         if (!forcePosStop) {
-          __leftHandlerPosition = __dAxis - __axisDragTmp; // - (_touchSize);
+          __leftHandlerPosition = __dAxis! - __axisDragTmp!; // - (_touchSize);
 
           _leftHandlerMoveBetweenSteps(__leftHandlerPosition, selectedByTap);
-          tmpLowerValue = _lowerValue;
+          tmpLowerValue = _lowerValue!;
         } else {
-          if (__axisPosTmp - tS >= __rightHandlerPosition) {
+          if (__axisPosTmp! - tS! >= __rightHandlerPosition!) {
             __leftHandlerPosition = __rightHandlerPosition;
-            _lowerValue = tmpLowerValue = _upperValue;
+            _lowerValue = tmpLowerValue = _upperValue!;
           } else {
             __leftHandlerPosition = getPositionByValue(_realMin);
-            _lowerValue = tmpLowerValue = _realMin;
+            _lowerValue = tmpLowerValue = _realMin!;
           }
           _updateLowerValue(tmpLowerValue);
         }
@@ -1070,11 +1073,11 @@ class _FlutterSliderState extends State<FlutterSlider>
     }
   }
 
-  bool _leftHandlerIgnoreSteps(double tS) {
+  bool _leftHandlerIgnoreSteps(double? tS) {
     bool validMove = true;
     if (_ignoreSteps.length > 0) {
-      if (__axisPosTmp <= 0) {
-        double ignorePoint;
+      if (__axisPosTmp! <= 0) {
+        double? ignorePoint;
         if (widget.rtl)
           ignorePoint = _findBiggestIgnorePoint();
         else
@@ -1084,7 +1087,7 @@ class _FlutterSliderState extends State<FlutterSlider>
         _lowerValue = ignorePoint;
         _updateLowerValue(_lowerValue);
         return false;
-      } else if (__axisPosTmp - tS >= __rightHandlerPosition) {
+      } else if (__axisPosTmp! - tS! >= __rightHandlerPosition!) {
         __leftHandlerPosition = __rightHandlerPosition;
         _lowerValue = _upperValue;
         _updateLowerValue(_lowerValue);
@@ -1093,15 +1096,15 @@ class _FlutterSliderState extends State<FlutterSlider>
 
       for (FlutterSliderIgnoreSteps steps in _ignoreSteps) {
         if (((!widget.rtl) &&
-                ((getValueByPositionIgnoreOffset(__axisPosTmp) >
-                        steps.from - _widgetStep / 2 &&
-                    getValueByPositionIgnoreOffset(__axisPosTmp) <=
-                        steps.to + _widgetStep / 2))) ||
+                (getValueByPositionIgnoreOffset(__axisPosTmp!) >
+                        steps.from! - _widgetStep! / 2 &&
+                    getValueByPositionIgnoreOffset(__axisPosTmp!) <=
+                        steps.to! + _widgetStep! / 2)) ||
             ((widget.rtl) &&
-                ((_realMax - getValueByPositionIgnoreOffset(__axisPosTmp) >
-                        steps.from - _widgetStep / 2 &&
-                    _realMax - getValueByPositionIgnoreOffset(__axisPosTmp) <=
-                        steps.to + _widgetStep / 2)))) validMove = false;
+                (_realMax! - getValueByPositionIgnoreOffset(__axisPosTmp!) >
+                        steps.from! - _widgetStep! / 2 &&
+                    _realMax! - getValueByPositionIgnoreOffset(__axisPosTmp!) <=
+                        steps.to! + _widgetStep! / 2))) validMove = false;
       }
     }
 
@@ -1110,27 +1113,27 @@ class _FlutterSliderState extends State<FlutterSlider>
 
   void _leftHandlerMoveBetweenSteps(handlerPos, bool selectedByTap) {
     double nextStepMiddlePos =
-        getPositionByValue((_lowerValue + (_lowerValue + _widgetStep)) / 2);
+        getPositionByValue((_lowerValue! + (_lowerValue! + _widgetStep!)) / 2);
     double prevStepMiddlePos =
-        getPositionByValue((_lowerValue - (_lowerValue - _widgetStep)) / 2);
+        getPositionByValue((_lowerValue! - (_lowerValue! - _widgetStep!)) / 2);
 
     if (handlerPos > nextStepMiddlePos || handlerPos < prevStepMiddlePos) {
       if (handlerPos > nextStepMiddlePos) {
-        _lowerValue += _widgetStep;
-        if (_lowerValue > _realMax) _lowerValue = _realMax;
-        if (_lowerValue > _upperValue) _lowerValue = _upperValue;
+        _lowerValue = _lowerValue! + _widgetStep!;
+        if (_lowerValue! > _realMax!) _lowerValue = _realMax;
+        if (_lowerValue! > _upperValue!) _lowerValue = _upperValue;
       } else {
-        _lowerValue -= _widgetStep;
-        if (_lowerValue < _realMin) _lowerValue = _realMin;
+        _lowerValue = _lowerValue! - _widgetStep!;
+        if (_lowerValue! < _realMin!) _lowerValue = _realMin;
       }
     }
     _updateLowerValue(_lowerValue);
   }
 
   void _lockedHandlers(handler) {
-    double distanceOfTwoHandlers = getLengthByValue(_handlersDistance);
+    double? distanceOfTwoHandlers = getLengthByValue(_handlersDistance);
 
-    double leftHandlerPos, rightHandlerPos;
+    double? leftHandlerPos, rightHandlerPos;
     if (widget.axis == Axis.horizontal) {
       leftHandlerPos = _leftHandlerXPosition;
       rightHandlerPos = _rightHandlerXPosition;
@@ -1140,22 +1143,22 @@ class _FlutterSliderState extends State<FlutterSlider>
     }
 
     if (handler == 'rightHandler') {
-      _lowerValue = _upperValue - _handlersDistance;
-      leftHandlerPos = rightHandlerPos - distanceOfTwoHandlers;
-      if (getValueByPositionIgnoreOffset(__axisPosTmp) - _handlersDistance <
-          _realMin) {
+      _lowerValue = _upperValue! - _handlersDistance;
+      leftHandlerPos = rightHandlerPos! - distanceOfTwoHandlers!;
+      if (getValueByPositionIgnoreOffset(__axisPosTmp!) - _handlersDistance <
+          _realMin!) {
         _lowerValue = _realMin;
-        _upperValue = _realMin + _handlersDistance;
+        _upperValue = _realMin! + _handlersDistance;
         rightHandlerPos = getPositionByValue(_upperValue);
         leftHandlerPos = getPositionByValue(_lowerValue);
       }
     } else {
-      _upperValue = _lowerValue + _handlersDistance;
-      rightHandlerPos = leftHandlerPos + distanceOfTwoHandlers;
-      if (getValueByPositionIgnoreOffset(__axisPosTmp) + _handlersDistance >
-          _realMax) {
+      _upperValue = _lowerValue! + _handlersDistance;
+      rightHandlerPos = leftHandlerPos! + distanceOfTwoHandlers!;
+      if (getValueByPositionIgnoreOffset(__axisPosTmp!) + _handlersDistance >
+          _realMax!) {
         _upperValue = _realMax;
-        _lowerValue = _realMax - _handlersDistance;
+        _lowerValue = _realMax! - _handlersDistance;
         rightHandlerPos = getPositionByValue(_upperValue);
         leftHandlerPos = getPositionByValue(_lowerValue);
       }
@@ -1176,16 +1179,16 @@ class _FlutterSliderState extends State<FlutterSlider>
   void _updateLowerValue(value) {
     _outputLowerValue = _displayRealValue(value);
     if (widget.rtl == true) {
-      _outputLowerValue = _displayRealValue(_realMax - value);
+      _outputLowerValue = _displayRealValue(_realMax! - value);
     }
   }
 
   void _rightHandlerMove(PointerEvent pointer,
       {double tappedPositionWithPadding = 0, bool selectedByTap = false}) {
     if (widget.disabled ||
-        (widget.rightHandler != null && widget.rightHandler.disabled)) return;
+        (widget.rightHandler != null && widget.rightHandler!.disabled)) return;
 
-    _handlersDistance = widget.lockDistance ?? _upperValue - _lowerValue;
+    _handlersDistance = widget.lockDistance ?? _upperValue! - _lowerValue!;
 
     if (selectedByTap) {
       _callbacks('onDragStarted', 1);
@@ -1212,68 +1215,69 @@ class _FlutterSliderState extends State<FlutterSlider>
           _constraintMaxHeight - _handlersPadding + 1;
     }
 
-    __axisPosTmp = __dAxis - __axisDragTmp + (_touchSize);
+    __axisPosTmp = __dAxis! - __axisDragTmp! + _touchSize!;
 
-    _checkRangeStep(getValueByPositionIgnoreOffset(__axisPosTmp));
+    _checkRangeStep(getValueByPositionIgnoreOffset(__axisPosTmp!));
 
-    __rAxis = getValueByPosition(__axisPosTmp);
+    __rAxis = getValueByPosition(__axisPosTmp!);
 
     if (widget.rangeSlider &&
         widget.minimumDistance > 0 &&
-        (__rAxis - widget.minimumDistance) <= _lowerValue) {
-      _upperValue = (_lowerValue + widget.minimumDistance < _realMax)
-          ? _lowerValue + widget.minimumDistance
+        (__rAxis! - widget.minimumDistance) <= _lowerValue!) {
+      _upperValue = (_lowerValue! + widget.minimumDistance < _realMax!)
+          ? _lowerValue! + widget.minimumDistance
           : _realMax;
       validMove = validMove & false;
       _updateUpperValue(_upperValue);
     }
     if (widget.rangeSlider &&
         widget.maximumDistance > 0 &&
-        __rAxis >= (_lowerValue + widget.maximumDistance)) {
-      _upperValue = (_lowerValue + widget.maximumDistance < _realMax)
-          ? _lowerValue + widget.maximumDistance
+        __rAxis! >= (_lowerValue! + widget.maximumDistance)) {
+      _upperValue = (_lowerValue! + widget.maximumDistance < _realMax!)
+          ? _lowerValue! + widget.maximumDistance
           : _realMax;
       validMove = validMove & false;
       _updateUpperValue(_upperValue);
     }
 
-    double tS = _touchSize;
+    double? tS = _touchSize;
     double rM = _handlersPadding;
     if (widget.jump) {
-      rM = -_handlersWidth;
-      tS = -_touchSize;
+      rM = -_handlersWidth!;
+      tS = -_touchSize!;
     }
 
     validMove = validMove & _rightHandlerIgnoreSteps(tS);
 
     bool forcePosStop = false;
-    if (((__axisPosTmp >= __containerSizeWithoutPadding) ||
-        (__axisPosTmp - tS <= __leftHandlerPosition))) {
+    if (((__axisPosTmp! >= __containerSizeWithoutPadding!) ||
+        (__axisPosTmp! - tS! <= __leftHandlerPosition!))) {
       forcePosStop = true;
     }
 
     if (validMove &&
-        (__axisPosTmp + rM <= __containerSizeWithoutHalfPadding ||
+        (__axisPosTmp! + rM <= __containerSizeWithoutHalfPadding! ||
             forcePosStop)) {
-      double tmpUpperValue = __rAxis;
+      double tmpUpperValue = __rAxis!;
 
-      if (tmpUpperValue > _realMax) tmpUpperValue = _realMax;
-      if (tmpUpperValue < _realMin) tmpUpperValue = _realMin;
+      if (tmpUpperValue > _realMax!) tmpUpperValue = _realMax!;
+      if (tmpUpperValue < _realMin!) tmpUpperValue = _realMin!;
 
-      if (tmpUpperValue < _lowerValue) tmpUpperValue = _lowerValue;
+      if (tmpUpperValue < _lowerValue!) tmpUpperValue = _lowerValue!;
 
       if (widget.jump == true) {
         if (!forcePosStop) {
           _upperValue = tmpUpperValue;
-          _rightHandlerMoveBetweenSteps(__dAxis - __axisDragTmp, selectedByTap);
+          _rightHandlerMoveBetweenSteps(
+              __dAxis! - __axisDragTmp!, selectedByTap);
           __rightHandlerPosition = getPositionByValue(_upperValue);
         } else {
-          if (__axisPosTmp - tS <= __leftHandlerPosition) {
+          if (__axisPosTmp! - tS! <= __leftHandlerPosition!) {
             __rightHandlerPosition = __leftHandlerPosition;
-            _upperValue = tmpUpperValue = _lowerValue;
+            _upperValue = tmpUpperValue = _lowerValue!;
           } else {
             __rightHandlerPosition = getPositionByValue(_realMax);
-            _upperValue = tmpUpperValue = _realMax;
+            _upperValue = tmpUpperValue = _realMax!;
           }
 
           _updateUpperValue(tmpUpperValue);
@@ -1282,16 +1286,16 @@ class _FlutterSliderState extends State<FlutterSlider>
         _upperValue = tmpUpperValue;
 
         if (!forcePosStop) {
-          __rightHandlerPosition = __dAxis - __axisDragTmp;
+          __rightHandlerPosition = __dAxis! - __axisDragTmp!;
           _rightHandlerMoveBetweenSteps(__rightHandlerPosition, selectedByTap);
-          tmpUpperValue = _upperValue;
+          tmpUpperValue = _upperValue!;
         } else {
-          if (__axisPosTmp - tS <= __leftHandlerPosition) {
+          if (__axisPosTmp! - tS! <= __leftHandlerPosition!) {
             __rightHandlerPosition = __leftHandlerPosition;
-            _upperValue = tmpUpperValue = _lowerValue;
+            _upperValue = tmpUpperValue = _lowerValue!;
           } else {
             __rightHandlerPosition = getPositionByValue(_realMax) + 1;
-            _upperValue = tmpUpperValue = _realMax;
+            _upperValue = tmpUpperValue = _realMax!;
           }
         }
         _updateUpperValue(tmpUpperValue);
@@ -1317,12 +1321,12 @@ class _FlutterSliderState extends State<FlutterSlider>
     }
   }
 
-  bool _rightHandlerIgnoreSteps(double tS) {
+  bool _rightHandlerIgnoreSteps(double? tS) {
     bool validMove = true;
     if (_ignoreSteps.length > 0) {
-      if (__axisPosTmp <= 0) {
+      if (__axisPosTmp! <= 0) {
         if (!widget.rangeSlider) {
-          double ignorePoint;
+          double? ignorePoint;
           if (widget.rtl)
             ignorePoint = _findBiggestIgnorePoint();
           else
@@ -1337,8 +1341,8 @@ class _FlutterSliderState extends State<FlutterSlider>
           _updateUpperValue(_upperValue);
         }
         return false;
-      } else if (__axisPosTmp >= __containerSizeWithoutPadding) {
-        double ignorePoint;
+      } else if (__axisPosTmp! >= __containerSizeWithoutPadding!) {
+        double? ignorePoint;
 
         if (widget.rtl)
           ignorePoint = _findSmallestIgnorePoint();
@@ -1353,33 +1357,33 @@ class _FlutterSliderState extends State<FlutterSlider>
 
       for (FlutterSliderIgnoreSteps steps in _ignoreSteps) {
         if (((!widget.rtl) &&
-                ((getValueByPositionIgnoreOffset(__axisPosTmp) >
-                        steps.from - _widgetStep / 2 &&
-                    getValueByPositionIgnoreOffset(__axisPosTmp) <=
-                        steps.to + _widgetStep / 2))) ||
+                (getValueByPositionIgnoreOffset(__axisPosTmp!) >
+                        steps.from! - _widgetStep! / 2 &&
+                    getValueByPositionIgnoreOffset(__axisPosTmp!) <=
+                        steps.to! + _widgetStep! / 2)) ||
             ((widget.rtl) &&
-                ((_realMax - getValueByPositionIgnoreOffset(__axisPosTmp) >
-                        steps.from - _widgetStep / 2 &&
-                    _realMax - getValueByPositionIgnoreOffset(__axisPosTmp) <=
-                        steps.to + _widgetStep / 2)))) validMove = false;
+                (_realMax! - getValueByPositionIgnoreOffset(__axisPosTmp!) >
+                        steps.from! - _widgetStep! / 2 &&
+                    _realMax! - getValueByPositionIgnoreOffset(__axisPosTmp!) <=
+                        steps.to! + _widgetStep! / 2))) validMove = false;
       }
     }
     return validMove;
   }
 
-  double _findSmallestIgnorePoint({ignoreBeyondBoundaries = false}) {
-    double ignorePoint = _realMax;
+  double? _findSmallestIgnorePoint({ignoreBeyondBoundaries = false}) {
+    double? ignorePoint = _realMax;
     bool beyondBoundaries = false;
     for (FlutterSliderIgnoreSteps steps in _ignoreSteps) {
-      if (steps.from < _realMin) beyondBoundaries = true;
-      if (steps.from < ignorePoint && steps.from >= _realMin)
-        ignorePoint = steps.from - _widgetStep;
-      else if (steps.to < ignorePoint && steps.to >= _realMin)
-        ignorePoint = steps.to + _widgetStep;
+      if (steps.from! < _realMin!) beyondBoundaries = true;
+      if (steps.from! < ignorePoint! && steps.from! >= _realMin!)
+        ignorePoint = steps.from! - _widgetStep!;
+      else if (steps.to! < ignorePoint && steps.to! >= _realMin!)
+        ignorePoint = steps.to! + _widgetStep!;
     }
     if (beyondBoundaries || ignoreBeyondBoundaries) {
       if (widget.rtl) {
-        ignorePoint = _realMax - ignorePoint;
+        ignorePoint = _realMax! - ignorePoint!;
       }
       return ignorePoint;
     } else {
@@ -1388,20 +1392,20 @@ class _FlutterSliderState extends State<FlutterSlider>
     }
   }
 
-  double _findBiggestIgnorePoint({ignoreBeyondBoundaries = false}) {
-    double ignorePoint = _realMin;
+  double? _findBiggestIgnorePoint({ignoreBeyondBoundaries = false}) {
+    double? ignorePoint = _realMin;
     bool beyondBoundaries = false;
     for (FlutterSliderIgnoreSteps steps in _ignoreSteps) {
-      if (steps.to > _realMax) beyondBoundaries = true;
+      if (steps.to! > _realMax!) beyondBoundaries = true;
 
-      if (steps.to > ignorePoint && steps.to <= _realMax)
-        ignorePoint = steps.to + _widgetStep;
-      else if (steps.from > ignorePoint && steps.from <= _realMax)
-        ignorePoint = steps.from - _widgetStep;
+      if (steps.to! > ignorePoint! && steps.to! <= _realMax!)
+        ignorePoint = steps.to! + _widgetStep!;
+      else if (steps.from! > ignorePoint && steps.from! <= _realMax!)
+        ignorePoint = steps.from! - _widgetStep!;
     }
     if (beyondBoundaries || ignoreBeyondBoundaries) {
       if (widget.rtl) {
-        ignorePoint = _realMax - ignorePoint;
+        ignorePoint = _realMax! - ignorePoint!;
       }
 
       return ignorePoint;
@@ -1413,18 +1417,18 @@ class _FlutterSliderState extends State<FlutterSlider>
 
   void _rightHandlerMoveBetweenSteps(handlerPos, bool selectedByTap) {
     double nextStepMiddlePos =
-        getPositionByValue((_upperValue + (_upperValue + _widgetStep)) / 2);
+        getPositionByValue((_upperValue! + (_upperValue! + _widgetStep!)) / 2);
     double prevStepMiddlePos =
-        getPositionByValue((_upperValue - (_upperValue - _widgetStep)) / 2);
+        getPositionByValue((_upperValue! - (_upperValue! - _widgetStep!)) / 2);
 
     if (handlerPos > nextStepMiddlePos || handlerPos < prevStepMiddlePos) {
       if (handlerPos > nextStepMiddlePos) {
-        _upperValue += _widgetStep;
-        if (_upperValue > _realMax) _upperValue = _realMax;
+        _upperValue = _upperValue! + _widgetStep!;
+        if (_upperValue! > _realMax!) _upperValue = _realMax;
       } else {
-        _upperValue -= _widgetStep;
-        if (_upperValue < _realMin) _upperValue = _realMin;
-        if (_upperValue < _lowerValue) _upperValue = _lowerValue;
+        _upperValue = _upperValue! - _widgetStep!;
+        if (_upperValue! < _realMin!) _upperValue = _realMin;
+        if (_upperValue! < _lowerValue!) _upperValue = _lowerValue;
       }
     }
     _updateUpperValue(_upperValue);
@@ -1433,23 +1437,23 @@ class _FlutterSliderState extends State<FlutterSlider>
   void _updateUpperValue(value) {
     _outputUpperValue = _displayRealValue(value);
     if (widget.rtl == true) {
-      _outputUpperValue = _displayRealValue(_realMax - value);
+      _outputUpperValue = _displayRealValue(_realMax! - value);
     }
   }
 
   void _checkRangeStep(double realValue) {
-    double sliderFromRange, sliderToRange;
+    double? sliderFromRange, sliderToRange;
     if (widget.step.rangeList != null) {
-      for (FlutterSliderRangeStep rangeStep in widget.step.rangeList) {
+      for (FlutterSliderRangeStep rangeStep in widget.step.rangeList!) {
         if (widget.step.isPercentRange) {
-          sliderFromRange = _widgetMax * rangeStep.from / 100;
-          sliderToRange = _widgetMax * rangeStep.to / 100;
+          sliderFromRange = _widgetMax! * rangeStep.from! / 100;
+          sliderToRange = _widgetMax! * rangeStep.to! / 100;
         } else {
           sliderFromRange = rangeStep.from;
           sliderToRange = rangeStep.to;
         }
 
-        if (realValue >= sliderFromRange && realValue <= sliderToRange) {
+        if (realValue >= sliderFromRange! && realValue <= sliderToRange!) {
           _widgetStep = rangeStep.step;
           _setDivisionAndDecimalScale();
           break;
@@ -1464,8 +1468,8 @@ class _FlutterSliderState extends State<FlutterSlider>
         child: Container(),
       );
 
-    double bottom;
-    double right;
+    double? bottom;
+    double? right;
     if (widget.axis == Axis.horizontal) {
       bottom = 0;
     } else {
@@ -1483,14 +1487,14 @@ class _FlutterSliderState extends State<FlutterSlider>
             axis: widget.axis,
             child: Stack(
               overflow: Overflow.visible,
-              children: <Widget>[
+              children: <Widget?>[
                 _tooltip(
                     side: 'left',
                     value: _outputLowerValue,
                     opacity: _leftTooltipOpacity,
                     animation: _leftTooltipAnimation),
                 leftHandler,
-              ],
+              ] as List<Widget>,
             ),
             feedback: Container()),
         onPointerMove: (_) {
@@ -1500,14 +1504,14 @@ class _FlutterSliderState extends State<FlutterSlider>
         },
         onPointerDown: (_) {
           if (widget.disabled ||
-              (widget.handler != null && widget.handler.disabled)) return;
+              (widget.handler != null && widget.handler!.disabled)) return;
 
           _renderBoxInitialization();
 
-          xDragTmp = (_.position.dx - _containerLeft - _leftHandlerXPosition);
-          yDragTmp = (_.position.dy - _containerTop - _leftHandlerYPosition);
+          xDragTmp = (_.position.dx - _containerLeft - _leftHandlerXPosition!);
+          yDragTmp = (_.position.dy - _containerTop - _leftHandlerYPosition!);
 
-          if (!_tooltipData.disabled &&
+          if (!_tooltipData.disabled! &&
               _tooltipData.alwaysShowTooltip == false) {
             _leftTooltipOpacity = 1;
             _leftTooltipAnimationController.forward();
@@ -1518,7 +1522,7 @@ class _FlutterSliderState extends State<FlutterSlider>
             }
           }
 
-          _leftHandlerScaleAnimationController.forward();
+          _leftHandlerScaleAnimationController!.forward();
 
           setState(() {});
 
@@ -1530,7 +1534,7 @@ class _FlutterSliderState extends State<FlutterSlider>
           _adjustLeftHandlerPosition();
 
           if (widget.disabled ||
-              (widget.handler != null && widget.handler.disabled)) return;
+              (widget.handler != null && widget.handler!.disabled)) return;
 
           _arrangeHandlersZIndex();
 
@@ -1552,22 +1556,22 @@ class _FlutterSliderState extends State<FlutterSlider>
     if (!widget.jump) {
       double position = getPositionByValue(_lowerValue);
       if (widget.axis == Axis.horizontal) {
-        _leftHandlerXPosition = position > _rightHandlerXPosition
+        _leftHandlerXPosition = position > _rightHandlerXPosition!
             ? _rightHandlerXPosition
             : position;
         if (widget.lockHandlers || __lockedHandlersDragOffset > 0) {
-          position = getPositionByValue(_lowerValue + _handlersDistance);
-          _rightHandlerXPosition = position < _leftHandlerXPosition
+          position = getPositionByValue(_lowerValue! + _handlersDistance);
+          _rightHandlerXPosition = position < _leftHandlerXPosition!
               ? _leftHandlerXPosition
               : position;
         }
       } else {
-        _leftHandlerYPosition = position > _rightHandlerYPosition
+        _leftHandlerYPosition = position > _rightHandlerYPosition!
             ? _rightHandlerYPosition
             : position;
         if (widget.lockHandlers || __lockedHandlersDragOffset > 0) {
-          position = getPositionByValue(_lowerValue + _handlersDistance);
-          _rightHandlerYPosition = position < _leftHandlerYPosition
+          position = getPositionByValue(_lowerValue! + _handlersDistance);
+          _rightHandlerYPosition = position < _leftHandlerYPosition!
               ? _leftHandlerYPosition
               : position;
         }
@@ -1576,7 +1580,7 @@ class _FlutterSliderState extends State<FlutterSlider>
   }
 
   void _hideTooltips() {
-    if (!_tooltipData.alwaysShowTooltip) {
+    if (!_tooltipData.alwaysShowTooltip!) {
       _leftTooltipOpacity = 0;
       _rightTooltipOpacity = 0;
       _leftTooltipAnimationController.reset();
@@ -1585,8 +1589,8 @@ class _FlutterSliderState extends State<FlutterSlider>
   }
 
   Positioned _rightHandlerWidget() {
-    double bottom;
-    double right;
+    double? bottom;
+    double? right;
     if (widget.axis == Axis.horizontal) {
       bottom = 0;
     } else {
@@ -1604,14 +1608,14 @@ class _FlutterSliderState extends State<FlutterSlider>
             axis: Axis.horizontal,
             child: Stack(
               overflow: Overflow.visible,
-              children: <Widget>[
+              children: <Widget?>[
                 _tooltip(
                     side: 'right',
                     value: _outputUpperValue,
                     opacity: _rightTooltipOpacity,
                     animation: _rightTooltipAnimation),
                 rightHandler,
-              ],
+              ] as List<Widget>,
             ),
             feedback: Container(
 //                            width: 20,
@@ -1621,7 +1625,7 @@ class _FlutterSliderState extends State<FlutterSlider>
         onPointerMove: (_) {
           __dragging = true;
 
-          if (!_tooltipData.disabled &&
+          if (!_tooltipData.disabled! &&
               _tooltipData.alwaysShowTooltip == false) {
             _rightTooltipOpacity = 1;
           }
@@ -1629,15 +1633,15 @@ class _FlutterSliderState extends State<FlutterSlider>
         },
         onPointerDown: (_) {
           if (widget.disabled ||
-              (widget.rightHandler != null && widget.rightHandler.disabled))
+              (widget.rightHandler != null && widget.rightHandler!.disabled))
             return;
 
           _renderBoxInitialization();
 
-          xDragTmp = (_.position.dx - _containerLeft - _rightHandlerXPosition);
-          yDragTmp = (_.position.dy - _containerTop - _rightHandlerYPosition);
+          xDragTmp = (_.position.dx - _containerLeft - _rightHandlerXPosition!);
+          yDragTmp = (_.position.dy - _containerTop - _rightHandlerYPosition!);
 
-          if (!_tooltipData.disabled &&
+          if (!_tooltipData.disabled! &&
               _tooltipData.alwaysShowTooltip == false) {
             _rightTooltipOpacity = 1;
             _rightTooltipAnimationController.forward();
@@ -1650,9 +1654,9 @@ class _FlutterSliderState extends State<FlutterSlider>
             setState(() {});
           }
           if (widget.rangeSlider == false)
-            _leftHandlerScaleAnimationController.forward();
+            _leftHandlerScaleAnimationController!.forward();
           else
-            _rightHandlerScaleAnimationController.forward();
+            _rightHandlerScaleAnimationController!.forward();
 
           _callbacks('onDragStarted', 1);
         },
@@ -1662,7 +1666,7 @@ class _FlutterSliderState extends State<FlutterSlider>
           _adjustRightHandlerPosition();
 
           if (widget.disabled ||
-              (widget.rightHandler != null && widget.rightHandler.disabled))
+              (widget.rightHandler != null && widget.rightHandler!.disabled))
             return;
 
           _arrangeHandlersZIndex();
@@ -1691,20 +1695,22 @@ class _FlutterSliderState extends State<FlutterSlider>
     if (!widget.jump) {
       double position = getPositionByValue(_upperValue);
       if (widget.axis == Axis.horizontal) {
-        _rightHandlerXPosition =
-            position < _leftHandlerXPosition ? _leftHandlerXPosition : position;
+        _rightHandlerXPosition = position < _leftHandlerXPosition!
+            ? _leftHandlerXPosition
+            : position;
         if (widget.lockHandlers) {
-          position = getPositionByValue(_upperValue - _handlersDistance);
-          _leftHandlerXPosition = position > _rightHandlerXPosition
+          position = getPositionByValue(_upperValue! - _handlersDistance);
+          _leftHandlerXPosition = position > _rightHandlerXPosition!
               ? _rightHandlerXPosition
               : position;
         }
       } else {
-        _rightHandlerYPosition =
-            position < _leftHandlerYPosition ? _leftHandlerYPosition : position;
+        _rightHandlerYPosition = position < _leftHandlerYPosition!
+            ? _leftHandlerYPosition
+            : position;
         if (widget.lockHandlers) {
-          position = getPositionByValue(_upperValue - _handlersDistance);
-          _leftHandlerYPosition = position > _rightHandlerYPosition
+          position = getPositionByValue(_upperValue! - _handlersDistance);
+          _leftHandlerYPosition = position > _rightHandlerYPosition!
               ? _rightHandlerYPosition
               : position;
         }
@@ -1713,19 +1719,19 @@ class _FlutterSliderState extends State<FlutterSlider>
   }
 
   void _stopHandlerAnimation(
-      {Animation animation, AnimationController controller}) {
+      {Animation? animation, AnimationController? controller}) {
     if (widget.handlerAnimation.reverseCurve != null) {
-      if (animation.isCompleted)
-        controller.reverse();
+      if (animation!.isCompleted)
+        controller!.reverse();
       else {
-        controller.reset();
+        controller!.reset();
       }
     } else
-      controller.reset();
+      controller!.reset();
   }
 
   drawHandlers() {
-    List<Positioned> items = []..addAll([
+    List<Positioned?> items = []..addAll([
         Function.apply(_inactiveTrack, []),
         Function.apply(_centralWidget, []),
         Function.apply(_activeTrack, []),
@@ -1745,7 +1751,7 @@ class _FlutterSliderState extends State<FlutterSlider>
             onPointerUp: (_) {
               if (widget.selectByTap && !__dragging) {
                 tappedPositionWithPadding = _distance();
-                if (_distanceFromLeftHandler < _distanceFromRightHandler) {
+                if (_distanceFromLeftHandler! < _distanceFromRightHandler!) {
                   if (!widget.rangeSlider) {
                     _rightHandlerMove(_,
                         tappedPositionWithPadding: tappedPositionWithPadding,
@@ -1800,7 +1806,7 @@ class _FlutterSliderState extends State<FlutterSlider>
                 if (widget.rangeSlider) {
                   if (_leftTapAndSlide) {
                     _trackBarSlideCallDragStated(0);
-                    if (!_tooltipData.disabled &&
+                    if (!_tooltipData.disabled! &&
                         _tooltipData.alwaysShowTooltip == false) {
                       _leftTooltipOpacity = 1;
                       _leftTooltipAnimationController.forward();
@@ -1809,7 +1815,7 @@ class _FlutterSliderState extends State<FlutterSlider>
                         tappedPositionWithPadding: tappedPositionWithPadding);
                   } else {
                     _trackBarSlideCallDragStated(1);
-                    if (!_tooltipData.disabled &&
+                    if (!_tooltipData.disabled! &&
                         _tooltipData.alwaysShowTooltip == false) {
                       _rightTooltipOpacity = 1;
                       _rightTooltipAnimationController.forward();
@@ -1819,7 +1825,7 @@ class _FlutterSliderState extends State<FlutterSlider>
                   }
                 } else {
                   _trackBarSlideCallDragStated(1);
-                  if (!_tooltipData.disabled &&
+                  if (!_tooltipData.disabled! &&
                       _tooltipData.alwaysShowTooltip == false) {
                     _rightTooltipOpacity = 1;
                     _rightTooltipAnimationController.forward();
@@ -1838,13 +1844,13 @@ class _FlutterSliderState extends State<FlutterSlider>
 
               double leftHandlerLastPosition, rightHandlerLastPosition;
               if (widget.axis == Axis.horizontal) {
-                double lX = _leftHandlerXPosition +
+                double lX = _leftHandlerXPosition! +
                     _handlersPadding +
-                    (_touchSize) +
+                    _touchSize! +
                     _containerLeft;
-                double rX = _rightHandlerXPosition +
+                double rX = _rightHandlerXPosition! +
                     _handlersPadding +
-                    (_touchSize) +
+                    _touchSize! +
                     _containerLeft;
 
                 _distanceFromRightHandler = (rX - _.position.dx);
@@ -1853,13 +1859,13 @@ class _FlutterSliderState extends State<FlutterSlider>
                 leftHandlerLastPosition = lX;
                 rightHandlerLastPosition = rX;
               } else {
-                double lY = _leftHandlerYPosition +
+                double lY = _leftHandlerYPosition! +
                     _handlersPadding +
-                    _touchSize +
+                    _touchSize! +
                     _containerTop;
-                double rY = _rightHandlerYPosition +
+                double rY = _rightHandlerYPosition! +
                     _handlersPadding +
-                    _touchSize +
+                    _touchSize! +
                     _containerTop;
 
                 _distanceFromLeftHandler = (lY - _.position.dy);
@@ -1872,21 +1878,21 @@ class _FlutterSliderState extends State<FlutterSlider>
               if (widget.rangeSlider &&
                   widget.trackBar.activeTrackBarDraggable &&
                   _ignoreSteps.length == 0 &&
-                  _distanceFromRightHandler > 0 &&
-                  _distanceFromLeftHandler < 0) {
+                  _distanceFromRightHandler! > 0 &&
+                  _distanceFromLeftHandler! < 0) {
                 _slidingByActiveTrackBar = true;
               } else {
                 double thumbPosition = (widget.axis == Axis.vertical)
                     ? _.position.dy
                     : _.position.dx;
-                if (_distanceFromLeftHandler.abs() <
-                        _distanceFromRightHandler.abs() ||
+                if (_distanceFromLeftHandler!.abs() <
+                        _distanceFromRightHandler!.abs() ||
                     (_distanceFromLeftHandler == _distanceFromRightHandler &&
                         thumbPosition < leftHandlerLastPosition)) {
                   _leftTapAndSlide = true;
                 }
-                if (_distanceFromRightHandler.abs() <
-                        _distanceFromLeftHandler.abs() ||
+                if (_distanceFromRightHandler!.abs() <
+                        _distanceFromLeftHandler!.abs() ||
                     (_distanceFromLeftHandler == _distanceFromRightHandler &&
                         thumbPosition < rightHandlerLastPosition)) {
                   _rightTapAndSlide = true;
@@ -1894,17 +1900,17 @@ class _FlutterSliderState extends State<FlutterSlider>
               }
 
               // if drag is within active area
-              if (_distanceFromRightHandler > 0 &&
-                  _distanceFromLeftHandler < 0) {
+              if (_distanceFromRightHandler! > 0 &&
+                  _distanceFromLeftHandler! < 0) {
                 if (widget.axis == Axis.horizontal) {
                   xDragTmp = 0;
                   __lockedHandlersDragOffset =
-                      ((_leftHandlerXPosition) + _containerLeft - _.position.dx)
+                      (_leftHandlerXPosition! + _containerLeft - _.position.dx)
                           .abs();
                 } else {
                   yDragTmp = 0;
                   __lockedHandlersDragOffset =
-                      ((_leftHandlerYPosition) + _containerTop - _.position.dy)
+                      (_leftHandlerYPosition! + _containerTop - _.position.dy)
                           .abs();
                 }
               }
@@ -1912,7 +1918,7 @@ class _FlutterSliderState extends State<FlutterSlider>
 
               if (_ignoreSteps.length == 0) {
                 if ((widget.lockHandlers || __lockedHandlersDragOffset > 0) &&
-                    !_tooltipData.disabled &&
+                    !_tooltipData.disabled! &&
                     _tooltipData.alwaysShowTooltip == false) {
                   _leftTooltipOpacity = 1;
                   _leftTooltipAnimationController.forward();
@@ -1921,8 +1927,8 @@ class _FlutterSliderState extends State<FlutterSlider>
                 }
 
                 if ((widget.lockHandlers || __lockedHandlersDragOffset > 0)) {
-                  _leftHandlerScaleAnimationController.forward();
-                  _rightHandlerScaleAnimationController.forward();
+                  _leftHandlerScaleAnimationController!.forward();
+                  _rightHandlerScaleAnimationController!.forward();
                 }
               }
 
@@ -1954,19 +1960,19 @@ class _FlutterSliderState extends State<FlutterSlider>
   }
 
   _distance() {
-    _distanceFromLeftHandler = _distanceFromLeftHandler.abs();
-    _distanceFromRightHandler = _distanceFromRightHandler.abs();
+    _distanceFromLeftHandler = _distanceFromLeftHandler!.abs();
+    _distanceFromRightHandler = _distanceFromRightHandler!.abs();
 
     if (widget.axis == Axis.horizontal) {
-      return _handlersWidth / 2 + (_touchSize) - xDragTmp;
+      return _handlersWidth! / 2 + _touchSize! - xDragTmp;
     } else {
-      return _handlersHeight / 2 + (_touchSize) - yDragTmp;
+      return _handlersHeight! / 2 + _touchSize! - yDragTmp;
     }
   }
 
   Positioned _tooltip(
-      {String side, dynamic value, double opacity, Animation animation}) {
-    if (_tooltipData.disabled || value == '')
+      {String? side, dynamic value, double? opacity, Animation? animation}) {
+    if (_tooltipData.disabled! || value == '')
       return Positioned(
         child: Container(),
       );
@@ -1987,7 +1993,7 @@ class _FlutterSliderState extends State<FlutterSlider>
     }
     String numberFormat = value.toString();
     if (_tooltipData.format != null)
-      numberFormat = _tooltipData.format(numberFormat);
+      numberFormat = _tooltipData.format!(numberFormat);
 
     List<Widget> children = [
       prefix,
@@ -2014,20 +2020,20 @@ class _FlutterSliderState extends State<FlutterSlider>
 //          height: __tooltipKEY.currentContext.size.height,
           key: (side == 'left') ? leftTooltipKey : rightTooltipKey,
 //            alignment: Alignment.center,
-          child: (widget.tooltip != null && widget.tooltip.custom != null)
-              ? widget.tooltip.custom(value)
+          child: (widget.tooltip != null && widget.tooltip!.custom != null)
+              ? widget.tooltip!.custom!(value)
               : Container(
                   padding: EdgeInsets.all(8),
-                  decoration: _tooltipData.boxStyle.decoration,
+                  decoration: _tooltipData.boxStyle!.decoration,
                   foregroundDecoration:
-                      _tooltipData.boxStyle.foregroundDecoration,
-                  transform: _tooltipData.boxStyle.transform,
+                      _tooltipData.boxStyle!.foregroundDecoration,
+                  transform: _tooltipData.boxStyle!.transform,
                   child: _tooltipHolderWidget),
         ),
       ),
     ));
 
-    double top, right, bottom, left;
+    double? top, right, bottom, left;
     switch (_tooltipData.direction) {
       case FlutterSliderTooltipDirection.top:
         top = 0;
@@ -2047,17 +2053,18 @@ class _FlutterSliderState extends State<FlutterSlider>
     }
 
     if (_tooltipData.positionOffset != null) {
-      if (_tooltipData.positionOffset.top != null)
-        top += _tooltipData.positionOffset.top;
-      if (_tooltipData.positionOffset.left != null)
-        left += _tooltipData.positionOffset.left;
-      if (_tooltipData.positionOffset.right != null)
-        right += _tooltipData.positionOffset.right;
-      if (_tooltipData.positionOffset.bottom != null)
-        bottom += _tooltipData.positionOffset.bottom;
+      if (_tooltipData.positionOffset!.top != null)
+        top = top! + _tooltipData.positionOffset!.top!;
+      if (_tooltipData.positionOffset!.left != null)
+        left = left! + _tooltipData.positionOffset!.left!;
+      if (_tooltipData.positionOffset!.right != null)
+        right = right! + _tooltipData.positionOffset!.right!;
+      if (_tooltipData.positionOffset!.bottom != null)
+        bottom = bottom! + _tooltipData.positionOffset!.bottom!;
     }
 
-    tooltipWidget = SlideTransition(position: animation, child: tooltipWidget);
+    tooltipWidget = SlideTransition(
+        position: animation as Animation<Offset>, child: tooltipWidget);
 
     return Positioned(
       left: left,
@@ -2065,7 +2072,7 @@ class _FlutterSliderState extends State<FlutterSlider>
       top: top,
       bottom: bottom,
       child: Opacity(
-        opacity: opacity,
+        opacity: opacity!,
         child: Center(child: tooltipWidget),
       ),
     );
@@ -2079,7 +2086,7 @@ class _FlutterSliderState extends State<FlutterSlider>
     if (widget.disabled)
       trackBarColor = widget.trackBar.inactiveDisabledTrackBarColor;
 
-    double top, bottom, left, right, width, height;
+    double? top, bottom, left, right, width, height;
     top = left = right = width = height = 0;
     right = bottom = null;
 
@@ -2127,7 +2134,7 @@ class _FlutterSliderState extends State<FlutterSlider>
     if (widget.disabled)
       trackBarColor = widget.trackBar.activeDisabledTrackBarColor;
 
-    double top, bottom, left, right, width, height;
+    double? top, bottom, left, right, width, height;
     top = left = width = height = 0;
     right = bottom = null;
 
@@ -2135,28 +2142,28 @@ class _FlutterSliderState extends State<FlutterSlider>
       bottom = 0;
       height = widget.trackBar.activeTrackBarHeight;
       if (!widget.centeredOrigin || widget.rangeSlider) {
-        width = _rightHandlerXPosition - _leftHandlerXPosition;
-        left = _leftHandlerXPosition + _handlersWidth / 2 + (_touchSize);
+        width = _rightHandlerXPosition! - _leftHandlerXPosition!;
+        left = _leftHandlerXPosition! + _handlersWidth! / 2 + _touchSize!;
 
         if (widget.rtl == true && widget.rangeSlider == false) {
           left = null;
-          right = _handlersWidth / 2;
-          width = _containerWidthWithoutPadding -
-              _rightHandlerXPosition -
-              _touchSize;
+          right = _handlersWidth! / 2;
+          width = _containerWidthWithoutPadding! -
+              _rightHandlerXPosition! -
+              _touchSize!;
         }
       } else {
-        if (_containerWidthWithoutPadding / 2 - _touchSize >
-            _rightHandlerXPosition) {
-          width = _containerWidthWithoutPadding / 2 -
-              _rightHandlerXPosition -
-              _touchSize;
-          left = _rightHandlerXPosition + _handlersWidth / 2 + (_touchSize);
+        if (_containerWidthWithoutPadding! / 2 - _touchSize! >
+            _rightHandlerXPosition!) {
+          width = _containerWidthWithoutPadding! / 2 -
+              _rightHandlerXPosition! -
+              _touchSize!;
+          left = _rightHandlerXPosition! + _handlersWidth! / 2 + _touchSize!;
         } else {
-          left = _containerWidthWithoutPadding / 2 + _handlersPadding;
-          width = _rightHandlerXPosition +
-              _touchSize -
-              _containerWidthWithoutPadding / 2;
+          left = _containerWidthWithoutPadding! / 2 + _handlersPadding;
+          width = _rightHandlerXPosition! +
+              _touchSize! -
+              _containerWidthWithoutPadding! / 2;
         }
       }
     } else {
@@ -2164,27 +2171,27 @@ class _FlutterSliderState extends State<FlutterSlider>
       width = widget.trackBar.activeTrackBarHeight;
 
       if (!widget.centeredOrigin || widget.rangeSlider) {
-        height = _rightHandlerYPosition - _leftHandlerYPosition;
-        top = _leftHandlerYPosition + _handlersHeight / 2 + (_touchSize);
+        height = _rightHandlerYPosition! - _leftHandlerYPosition!;
+        top = _leftHandlerYPosition! + _handlersHeight! / 2 + _touchSize!;
         if (widget.rtl == true && widget.rangeSlider == false) {
           top = null;
-          bottom = _handlersHeight / 2;
-          height = _containerHeightWithoutPadding -
-              _rightHandlerYPosition -
-              _touchSize;
+          bottom = _handlersHeight! / 2;
+          height = _containerHeightWithoutPadding! -
+              _rightHandlerYPosition! -
+              _touchSize!;
         }
       } else {
-        if (_containerHeightWithoutPadding / 2 - _touchSize >
-            _rightHandlerYPosition) {
-          height = _containerHeightWithoutPadding / 2 -
-              _rightHandlerYPosition -
-              _touchSize;
-          top = _rightHandlerYPosition + _handlersHeight / 2 + (_touchSize);
+        if (_containerHeightWithoutPadding! / 2 - _touchSize! >
+            _rightHandlerYPosition!) {
+          height = _containerHeightWithoutPadding! / 2 -
+              _rightHandlerYPosition! -
+              _touchSize!;
+          top = _rightHandlerYPosition! + _handlersHeight! / 2 + _touchSize!;
         } else {
-          top = _containerHeightWithoutPadding / 2 + _handlersPadding;
-          height = _rightHandlerYPosition +
-              _touchSize -
-              _containerHeightWithoutPadding / 2;
+          top = _containerHeightWithoutPadding! / 2 + _handlersPadding;
+          height = _rightHandlerYPosition! +
+              _touchSize! -
+              _containerHeightWithoutPadding! / 2;
         }
       }
     }
@@ -2236,29 +2243,29 @@ class _FlutterSliderState extends State<FlutterSlider>
     switch (callbackName) {
       case 'onDragging':
         if (widget.onDragging != null)
-          widget.onDragging(handlerIndex, lowerValue, upperValue);
+          widget.onDragging!(handlerIndex, lowerValue, upperValue);
         break;
       case 'onDragCompleted':
         if (widget.onDragCompleted != null)
-          widget.onDragCompleted(handlerIndex, lowerValue, upperValue);
+          widget.onDragCompleted!(handlerIndex, lowerValue, upperValue);
         break;
       case 'onDragStarted':
         if (widget.onDragStarted != null)
-          widget.onDragStarted(handlerIndex, lowerValue, upperValue);
+          widget.onDragStarted!(handlerIndex, lowerValue, upperValue);
         break;
     }
   }
 
-  dynamic _displayRealValue(double value) {
+  dynamic _displayRealValue(double? value) {
     if (_fixedValues.length > 0) {
-      return _fixedValues[value.toInt()].value;
+      return _fixedValues[value!.toInt()].value;
     }
 
-    return double.parse((value + _widgetMin).toStringAsFixed(_decimalScale));
+    return double.parse((value! + _widgetMin!).toStringAsFixed(_decimalScale));
   }
 
   void _arrangeHandlersZIndex() {
-    if (_lowerValue >= (_realMax / 2))
+    if (_lowerValue! >= (_realMax! / 2))
       _positionedItems = [
         _rightHandlerWidget,
         _leftHandlerWidget,
@@ -2275,31 +2282,31 @@ class _FlutterSliderState extends State<FlutterSlider>
         (MediaQuery.of(context).size.width - _constraintMaxWidth) <=
             _containerLeft) {
       RenderBox containerRenderBox =
-          containerKey.currentContext.findRenderObject();
+          containerKey.currentContext!.findRenderObject() as RenderBox;
       _containerLeft = containerRenderBox.localToGlobal(Offset.zero).dx;
     }
     if (_containerTop <= 0 ||
         (MediaQuery.of(context).size.height - _constraintMaxHeight) <=
             _containerTop) {
       RenderBox containerRenderBox =
-          containerKey.currentContext.findRenderObject();
+          containerKey.currentContext!.findRenderObject() as RenderBox;
       _containerTop = containerRenderBox.localToGlobal(Offset.zero).dy;
     }
   }
 }
 
 class _MakeHandler extends StatelessWidget {
-  final double width;
-  final double height;
-  final GlobalKey id;
-  final FlutterSliderHandler handlerData;
-  final bool visibleTouchArea;
-  final Animation animation;
-  final Axis axis;
-  final int handlerIndex;
+  final double? width;
+  final double? height;
+  final GlobalKey? id;
+  final FlutterSliderHandler? handlerData;
+  final bool? visibleTouchArea;
+  final Animation? animation;
+  final Axis? axis;
+  final int? handlerIndex;
   final bool rtl;
   final bool rangeSlider;
-  final double touchSize;
+  final double? touchSize;
 
   _MakeHandler(
       {this.id,
@@ -2319,8 +2326,8 @@ class _MakeHandler extends StatelessWidget {
     double touchOpacity = (visibleTouchArea == true) ? 1 : 0;
 
     double localWidth, localHeight;
-    localHeight = height + (touchSize * 2);
-    localWidth = width + (touchSize * 2);
+    localHeight = height! + (touchSize! * 2);
+    localWidth = width! + (touchSize! * 2);
 
     FlutterSliderHandler handler = handlerData ?? FlutterSliderHandler();
 
@@ -2362,7 +2369,7 @@ class _MakeHandler extends StatelessWidget {
           ),
           Center(
             child: ScaleTransition(
-              scale: animation,
+              scale: animation as Animation<double>,
               child: Opacity(
                 opacity: handler.opacity ?? 1,
                 child: Container(
@@ -2384,10 +2391,10 @@ class _MakeHandler extends StatelessWidget {
 }
 
 class FlutterSliderHandler {
-  BoxDecoration decoration;
-  BoxDecoration foregroundDecoration;
-  Matrix4 transform;
-  Widget child;
+  BoxDecoration? decoration;
+  BoxDecoration? foregroundDecoration;
+  Matrix4? transform;
+  Widget? child;
   bool disabled;
   double opacity;
 
@@ -2417,19 +2424,19 @@ class FlutterSliderHandler {
 }
 
 class FlutterSliderTooltip {
-  Widget Function(dynamic value) custom;
-  String Function(String value) format;
-  TextStyle textStyle;
-  FlutterSliderTooltipBox boxStyle;
-  Widget leftPrefix;
-  Widget leftSuffix;
-  Widget rightPrefix;
-  Widget rightSuffix;
-  bool alwaysShowTooltip;
-  bool disabled;
-  bool disableAnimation;
-  FlutterSliderTooltipDirection direction;
-  FlutterSliderTooltipPositionOffset positionOffset;
+  Widget Function(dynamic value)? custom;
+  String Function(String value)? format;
+  TextStyle? textStyle;
+  FlutterSliderTooltipBox? boxStyle;
+  Widget? leftPrefix;
+  Widget? leftSuffix;
+  Widget? rightPrefix;
+  Widget? rightSuffix;
+  bool? alwaysShowTooltip;
+  bool? disabled;
+  bool? disableAnimation;
+  FlutterSliderTooltipDirection? direction;
+  FlutterSliderTooltipPositionOffset? positionOffset;
 
   FlutterSliderTooltip({
     this.custom,
@@ -2474,10 +2481,10 @@ class FlutterSliderTooltip {
 }
 
 class FlutterSliderTooltipPositionOffset {
-  double top;
-  double left;
-  double right;
-  double bottom;
+  double? top;
+  double? left;
+  double? right;
+  double? bottom;
 
   FlutterSliderTooltipPositionOffset(
       {this.top, this.left, this.right, this.bottom});
@@ -2495,9 +2502,9 @@ class FlutterSliderTooltipPositionOffset {
 }
 
 class FlutterSliderTooltipBox {
-  final BoxDecoration decoration;
-  final BoxDecoration foregroundDecoration;
-  final Matrix4 transform;
+  final BoxDecoration? decoration;
+  final BoxDecoration? foregroundDecoration;
+  final Matrix4? transform;
 
   const FlutterSliderTooltipBox(
       {this.decoration, this.foregroundDecoration, this.transform});
@@ -2513,13 +2520,13 @@ class FlutterSliderTooltipBox {
 }
 
 class FlutterSliderTrackBar {
-  final BoxDecoration inactiveTrackBar;
-  final BoxDecoration activeTrackBar;
+  final BoxDecoration? inactiveTrackBar;
+  final BoxDecoration? activeTrackBar;
   final Color activeDisabledTrackBarColor;
   final Color inactiveDisabledTrackBarColor;
   final double activeTrackBarHeight;
   final double inactiveTrackBarHeight;
-  final Widget centralWidget;
+  final Widget? centralWidget;
   final bool activeTrackBarDraggable;
 
   const FlutterSliderTrackBar({
@@ -2557,8 +2564,8 @@ class FlutterSliderTrackBar {
 }
 
 class FlutterSliderIgnoreSteps {
-  final double from;
-  final double to;
+  final double? from;
+  final double? to;
 
   FlutterSliderIgnoreSteps({this.from, this.to})
       : assert(from != null && to != null && from <= to);
@@ -2570,7 +2577,7 @@ class FlutterSliderIgnoreSteps {
 }
 
 class FlutterSliderFixedValue {
-  final int percent;
+  final int? percent;
   final dynamic value;
 
   FlutterSliderFixedValue({this.percent, this.value})
@@ -2585,7 +2592,7 @@ class FlutterSliderFixedValue {
 
 class FlutterSliderHandlerAnimation {
   final Curve curve;
-  final Curve reverseCurve;
+  final Curve? reverseCurve;
   final Duration duration;
   final double scale;
 
@@ -2611,14 +2618,14 @@ class FlutterSliderHandlerAnimation {
 class FlutterSliderHatchMark {
   bool disabled;
   double density;
-  double linesDistanceFromTrackBar;
-  double labelsDistanceFromTrackBar;
-  List<FlutterSliderHatchMarkLabel> labels;
-  FlutterSliderSizedBox smallLine;
-  FlutterSliderSizedBox bigLine;
-  FlutterSliderSizedBox labelBox;
+  double? linesDistanceFromTrackBar;
+  double? labelsDistanceFromTrackBar;
+  List<FlutterSliderHatchMarkLabel>? labels;
+  FlutterSliderSizedBox? smallLine;
+  FlutterSliderSizedBox? bigLine;
+  FlutterSliderSizedBox? labelBox;
   FlutterSliderHatchMarkAlignment linesAlignment;
-  bool displayLines;
+  bool? displayLines;
 
   FlutterSliderHatchMark(
       {this.disabled = false,
@@ -2659,8 +2666,8 @@ class FlutterSliderHatchMark {
 }
 
 class FlutterSliderHatchMarkLabel {
-  final double percent;
-  final Widget label;
+  final double? percent;
+  final Widget? label;
 
   FlutterSliderHatchMarkLabel({
     this.percent,
@@ -2675,9 +2682,9 @@ class FlutterSliderHatchMarkLabel {
 }
 
 class FlutterSliderSizedBox {
-  final BoxDecoration decoration;
-  final BoxDecoration foregroundDecoration;
-  final Matrix4 transform;
+  final BoxDecoration? decoration;
+  final BoxDecoration? foregroundDecoration;
+  final Matrix4? transform;
   final double width;
   final double height;
 
@@ -2685,8 +2692,8 @@ class FlutterSliderSizedBox {
       {this.decoration,
       this.foregroundDecoration,
       this.transform,
-      @required this.height,
-      @required this.width})
+      required this.height,
+      required this.width})
       : assert(width != null && height != null && width > 0 && height > 0);
 
   @override
@@ -2706,7 +2713,7 @@ class FlutterSliderSizedBox {
 class FlutterSliderStep {
   final double step;
   final bool isPercentRange;
-  final List<FlutterSliderRangeStep> rangeList;
+  final List<FlutterSliderRangeStep>? rangeList;
 
   const FlutterSliderStep({
     this.step = 1,
@@ -2725,9 +2732,9 @@ class FlutterSliderStep {
 }
 
 class FlutterSliderRangeStep {
-  final double from;
-  final double to;
-  final double step;
+  final double? from;
+  final double? to;
+  final double? step;
 
   FlutterSliderRangeStep({
     this.from,
